tradeding_api_server/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
mysql-connector-python==9.2.0
bcrypt==4.0.1
httpx==0.28.1
python-dotenv==1.1.0
websockets==15.0.1
email-validator==2.2.0